Creating DAL and Business Layer
