+++ TourPlanner.sln
@@ -5,7 +5,11 @@ VisualStudioVersion = 17.7.34024.191
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TourPlanner", "TourPlanner\TourPlanner.csproj", "{92128B4F-607E-4D20-9FF3-189C8F49E21F}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TourPlannerTest", "TourPlannerTest\TourPlannerTest.csproj", "{93492CA8-F7DA-464E-87A6-3B0A122CFF71}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TourPlannerTest", "TourPlannerTest\TourPlannerTest.csproj", "{93492CA8-F7DA-464E-87A6-3B0A122CFF71}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TourPlannerDAL", "TourPlannerDAL\TourPlannerDAL.csproj", "{9D5469A4-E2DD-4103-A8FA-5B828B4C3A37}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TourPlannerBusinessLayer", "TourPlannerBusinessLayer\TourPlannerBusinessLayer.csproj", "{6D21F8A4-5042-4B31-B281-9A9277587A72}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -21,6 +25,14 @@ Global
 		{93492CA8-F7DA-464E-87A6-3B0A122CFF71}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{93492CA8-F7DA-464E-87A6-3B0A122CFF71}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{93492CA8-F7DA-464E-87A6-3B0A122CFF71}.Release|Any CPU.Build.0 = Release|Any CPU
+		{9D5469A4-E2DD-4103-A8FA-5B828B4C3A37}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{9D5469A4-E2DD-4103-A8FA-5B828B4C3A37}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{9D5469A4-E2DD-4103-A8FA-5B828B4C3A37}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{9D5469A4-E2DD-4103-A8FA-5B828B4C3A37}.Release|Any CPU.Build.0 = Release|Any CPU
+		{6D21F8A4-5042-4B31-B281-9A9277587A72}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{6D21F8A4-5042-4B31-B281-9A9277587A72}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{6D21F8A4-5042-4B31-B281-9A9277587A72}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{6D21F8A4-5042-4B31-B281-9A9277587A72}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
