+++ TourPlanner/ViewModels/TourLogViewModel.cs
@@ -1,12 +1,8 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Collections.ObjectModel;
+﻿using System.Collections.ObjectModel;
 using System.ComponentModel;
 using System.Diagnostics;
 using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using TourPlanner.Models;
+using TourPlannerBusinessLayer.Models;
 
 namespace TourPlanner.ViewModels
 {
