+++ TourPlanner/Views/AddTourWindow.xaml
@@ -1,4 +1,4 @@
-﻿<Window x:Class="TourPlanner.AddTourWindow"
+﻿<Window x:Class="TourPlanner.Views.AddTourWindow"
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         Title="Tours" Height="400" Width="400"
