+++ TourPlanner/Views/MainWindow.xaml
@@ -1,9 +1,10 @@
-﻿<Window x:Class="TourPlanner.MainWindow"
+﻿<Window x:Class="TourPlanner.Views.MainWindow"
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
         xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
         xmlns:local="clr-namespace:TourPlanner"
+        xmlns:subViews="clr-namespace:TourPlanner.Views.SubViews"
         mc:Ignorable="d"
         Title="Tour Planner" Height="600" Width="800"
         DataContext="{Binding RelativeSource={RelativeSource Self}}" Foreground="White" Style="{DynamicResource MyStyle}">
@@ -59,8 +60,8 @@
             <RowDefinition Height="*" />
         </Grid.RowDefinitions>
 
-        <local:UserMainMenu Grid.Row="0" Grid.ColumnSpan="3"/>
-        <local:UserSearchBar Grid.Row="1" Grid.ColumnSpan="3"/>
+        <subViews:UserMainMenu Grid.Row="0" Grid.ColumnSpan="3"/>
+        <subViews:UserSearchBar Grid.Row="1" Grid.ColumnSpan="3"/>
 
         <Border Grid.Row="2" Grid.Column="0" BorderBrush="Gainsboro" BorderThickness="1" Grid.RowSpan="2">
             <Grid>
