+++ TourPlanner/Views/MainWindow.xaml.cs
@@ -1,25 +1,11 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Collections.ObjectModel;
-using System.ComponentModel;
-using System.Diagnostics;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
+﻿using System.Diagnostics;
 using System.Windows;
 using System.Windows.Controls;
-using System.Windows.Data;
-using System.Windows.Documents;
 using System.Windows.Input;
-using System.Windows.Media;
-using System.Windows.Media.Imaging;
-using System.Windows.Navigation;
-using System.Windows.Shapes;
-using TourPlanner.Models;
 using TourPlanner.ViewModels;
+using TourPlannerBusinessLayer.Models;
 
-
-namespace TourPlanner
+namespace TourPlanner.Views
 {
     public partial class MainWindow : Window
     {
@@ -28,10 +14,7 @@ namespace TourPlanner
         public MainWindow()
         {
             InitializeComponent();
-
-            
             tourViewModel = new TourViewModel();
-
           //Setzen des Data Contexts
             DataContext = tourViewModel;
         }
