+++ TourPlanner/Views/SubViews/UserMainMenu.xaml
@@ -1,4 +1,4 @@
-﻿<UserControl x:Class="TourPlanner.UserMainMenu"
+﻿<UserControl x:Class="TourPlanner.Views.SubViews.UserMainMenu"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
