+++ TourPlanner/Views/SubViews/UserMainMenu.xaml.cs
@@ -1,0 +1,15 @@
+﻿using System.Windows.Controls;
+
+namespace TourPlanner.Views.SubViews
+{
+    /// <summary>
+    /// Interaktionslogik für UserMainMenu.xaml
+    /// </summary>
+    public partial class UserMainMenu : UserControl
+    {
+        public UserMainMenu()
+        {
+            InitializeComponent();
+        }
+    }
+}
