+++ TourPlanner/Views/SubViews/UserSearchBar.xaml
@@ -1,4 +1,4 @@
-﻿<UserControl x:Class="TourPlanner.UserSearchBar"
+﻿<UserControl x:Class="TourPlanner.Views.SubViews.UserSearchBar"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
