+++ TourPlanner/Views/SubViews/UserSearchBar.xaml.cs
@@ -1,0 +1,15 @@
+﻿using System.Windows.Controls;
+
+namespace TourPlanner.Views.SubViews
+{
+    /// <summary>
+    /// Interaktionslogik für UserSearchBar.xaml
+    /// </summary>
+    public partial class UserSearchBar : UserControl
+    {
+        public UserSearchBar()
+        {
+            InitializeComponent();
+        }
+    }
+}
