+++ TourPlannerBusinessLayer/Class1.cs
@@ -1,0 +1,7 @@
+﻿namespace TourPlannerBusinessLayer
+{
+    public class Class1
+    {
+
+    }
+}
