+++ TourPlannerDAL/Class1.cs
@@ -1,0 +1,7 @@
+﻿namespace TourPlannerDAL
+{
+    public class Class1
+    {
+
+    }
+}
